Refactor Spec.hs and split it up
